% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflicted.R
\docType{package}
\name{conflicted-package}
\alias{conflicted}
\alias{conflicted-package}
\title{conflicted: An Alternative Conflict Resolution Strategy}
\description{
R's default conflict management system gives the most recently
loaded package precedence. This can make it hard to detect conflicts,
particularly when introduced by an update to an existing package.
conflicted takes a different approach, making every conflict an error
and forcing you to choose which function to use.
}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
