% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflicts.R
\name{conflicts_declare}
\alias{conflicts_declare}
\title{Declare many conflict preferences at once}
\usage{
conflicts_declare(code)
}
\arguments{
\item{code}{A block of code that declares preferences.}
}
\description{
\code{\link[=conflict_prefer]{conflict_prefer()}} and friends register conflicts after each call. If you
are registering many conflicts this can be expensive
}
\examples{
conflicts_declare({
  conflict_prefer("filter", "dplyr")
  conflict_prefer_all("dplyr", "MASS")
})
}
