% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/favor.R
\name{conflicts_prefer}
\alias{conflicts_prefer}
\title{Declare many preferences at once}
\usage{
conflicts_prefer(..., .quiet = FALSE)
}
\arguments{
\item{...}{Functions to prefer in form \code{pkg::fun} or \code{pkg::fun()}.}

\item{.quiet}{If \code{TRUE}, all output will be suppressed}
}
\description{
This is a wrapper around \code{\link[=conflict_prefer]{conflict_prefer()}} that makes it easier to
declare many preferences at once with a convenient syntax.
}
\examples{
conflicts_prefer(
  dplyr::filter(),
  dplyr::lag(),
)
}
