% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shim-colon.R
\name{shim_colon}
\alias{shim_colon}
\title{Strict version of `:`}
\usage{
shim_colon(from, to)
}
\arguments{
\item{from, to}{Passed on to [seq()]}
}
\description{
This version of `:` throws an error if `from` is positive and `to` is
zero, to make the error more obvious if you've used `1:length(x)` instead
of [seq_along()]
}
\examples{
x <- numeric()
seq_along(x)
lax(1:length(x))
}
