% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shim-risky.R
\name{strict_arg}
\alias{strict_arg}
\title{Strict arguments}
\usage{
strict_arg(name)
}
\arguments{
\item{name}{Name of argument}
}
\description{
A number of functions in base R have defaults that depend on a global
option, or a complicated expression. These shims force you to specify
a value for these arguments
}
