% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{strict_deactivate}
\alias{strict_deactivate}
\alias{strict_activate}
\title{Manually activate and deactive strict mode}
\usage{
strict_deactivate()

strict_activate()
}
\description{
For experts only.
}
\keyword{internal}
