% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shim-scalar.R
\name{strict_sample}
\alias{strict_sample}
\alias{strict_diag}
\title{Strict behaviour for functions with special scalar behaviour}
\usage{
strict_sample(x, size = length(x), replace = FALSE, prob = NULL)

strict_diag(x = 1, nrow, ncol)
}
\arguments{
\item{x, size, replace, prob, nrow, ncol}{Passed on to [sample()] and [diag()]}
}
\description{
[sample()] and [diag()] behave differently depending on whether their
first argument is a scalar or a function. These shims throw an error
when given a scalar to force you to pick a safer alternative.
}
\examples{
lax({
  sample(5:3)
  sample(5:4)
  sample(5:5)

  diag(5:3)
  diag(5:4)
  diag(5:5)
})

\dontrun{
  sample(5:5)
  diag(5)
}
}
