% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{conflict_scout}
\alias{conflict_scout}
\title{Find conflicts amongst a set of packages}
\usage{
conflict_scout(pkgs = NULL)
}
\arguments{
\item{pkgs}{Set of packages for which to report conflicts. If \code{NULL},
the default, will report conflicts for all loaded packages}
}
\value{
A named list of character vectors. The names are functions and
the values are the packages where they appear. If there is only a single
package listed, it means that an automated disambiguation has selected
that function.

A user friendly print method displays the result as bulleted list.
}
\description{
\code{conflict_scout()} is the workhorse behind the conflicted package. You can
call it directly yourself if you want to see all conflicts before hitting
them in practice.
}
\examples{
conflict_scout()
}
