% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflicted.R
\docType{package}
\name{conflicted-package}
\alias{conflicted}
\alias{conflicted-package}
\title{conflicted: An Alternative Conflict Resolution Strategy}
\description{
R's default conflict management system gives the most recently
    loaded package precedence. This can make it hard to detect conflicts, 
    particularly when they arise because a package update creates ambiguity
    that did not previously exist. 'conflicted' takes a different approach, 
    making every conflict an error and forcing you to choose which function 
    to use.
}
\section{Resolving conflicts}{


To permanently resolve a conflict within a session, use assignment:

\preformatted{
library(conflicted)
library(dplyr)

filter <- dplyr::filter
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/conflicted}
  \item Report bugs at \url{https://github.com/r-lib/conflicted/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
